requirements.txt
arrow==0.15.6
asgiref==3.2.7
attrs==19.3.0
blessed==1.17.6
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.5
django-allauth==0.41.0
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-extensions==2.2.9
django-filter==2.2.0
django-jsonfield==1.4.0
django-mailer==2.0.1
django-oauth-toolkit==1.3.2
django-picklefield==2.1.1
django-q==1.2.3
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-camel-case==1.1.2
djangorestframework-recursive==0.1.2
drf-yasg==1.17.1
future==0.18.2
idna==2.9
importlib-metadata==1.6.0
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
jsonlines==1.2.0
jsonschema==3.2.0
lockfile==0.12.2
Markdown==3.2.1
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.3
Pillow==7.1.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
swagger-spec-validator==2.5.0
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.25.9
user-agents==2.1
wcwidth==0.2.2
zipp==3.1.0
